micromap::set

Struct Set

pub struct Set<T, const N: usize> { /* private fields */ }

A faster alternative of std::collections::HashSet.

For example, this is how you make a set, which is allocated on stack and is capable of storing up to eight key-values pairs:

use micromap::Set;
let mut set : Set<u64, 8> = Set::new();
set.insert(1);
set.insert(2);
assert_eq!(set.len(), 2);
assert_eq!(set.capacity(), 8);

It is faster because it doesn’t use a hash function at all. It simply keeps all items in an array and when it’s necessary to find a value, it goes through all items comparing the needle with each pair available. Also it is faster because it doesn’t use heap. When a Set is being created, it allocates the necessary space on stack. That’s why the maximum size of the set must be provided in compile time.

It is also faster because it doesn’t grow in size. When a Set is created, its size is fixed on stack (or as Box<Set<_>> on heap if you like). If an attempt is made to insert() too many value into it, it simply panics (or you can use checked_insert()). Moreover, for method insert_unchecked() in the “release” mode, it doesn’t panic, instead, its behaviour is undefined. In the “release” mode all boundary checks are disabled, for the sake of higher performance (But in general, you can just use insert() or checked_insert() in most cases).

Implementations

impl<T, const N: usize> Set<T, N>

pub const fn new() -> Self

Creates an empty Set.

The set is initially created with a capacity of N, so even if you don’t insert any values, it will occupy a fixed stack memory space.

Examples

use micromap::Set;
let set: Set<i32, 8> = Set::new();
assert_eq!(set.len(), 0);
assert_eq!(set.capacity(), 8);

impl<'a, T: PartialEq + ?Sized, const N: usize> Set<&'a T, N>

pub fn difference_ref<'b, 'set1, 'set2, const M: usize>( &'set1 self, other: &'set2 Set<&'b T, M>, ) -> DifferenceRef<'a, 'b, 'set2, T, M>
where 'set1: 'a, 'set2: 'b,

A difference() method with more elaborate lifetime and just for Set<&T>.

impl<T: PartialEq, const N: usize> Set<T, N>

pub fn difference<'a, 'b, const M: usize>( &'a self, other: &'b Set<T, M>, ) -> Difference<'a, 'b, T, M>

Visits the values representing the difference, i.e., the values that are in self but not in other.

Examples

use micromap::Set;
let a = Set::from([1, 2, 3]);
let b = Set::from([4, 2, 3, 4]);
// Can be seen as `a - b`.
for x in a.difference(&b) {
    println!("{x}"); // Print 1
}
let diff: Set<_, 3> = a.difference(&b).copied().collect();
assert_eq!(diff, Set::from([1]));
// Note that difference is not symmetric,
// and `b - a` means something else:
let diff: Set<_, 4> = b.difference(&a).copied().collect();
assert_eq!(diff, Set::from([4]));

impl<T, const N: usize> Set<T, N>

pub fn drain(&mut self) -> Drain<'_, T>

Clears the set, returning all elements as an iterator. Keeps the allocated memory for reuse.

If the returned iterator is dropped before being fully consumed, it drops the remaining elements. The returned iterator keeps a mutable borrow on the set to optimize its implementation.

Examples

use micromap::Set;
let mut set = Set::from([1, 2, 3]);
assert!(!set.is_empty());
// print 1, 2, 3 in an arbitrary order
for i in set.drain() {
    println!("{i}");
}
assert!(set.is_empty());

impl<T: PartialEq, const N: usize> Set<T, N>

pub fn intersection<'a, const M: usize>( &'a self, other: &'a Set<T, M>, ) -> Intersection<'a, T, M>

Visits the values representing the intersection, i.e., the values that are both in self and other.

When an equal element is present in self and other, unlike the standard library functions, the resulting Intersection will ALWAYS yield references to the caller(self). This can be relevant if T contains fields which are not compared by its Eq implementation, and may hold different value between the two equal copies of T in the two sets.

Examples

use micromap::Set;
let a = Set::from([1, 2, 3]);
let b = Set::from([4, 2, 3, 4]);
// Print 2, 3 in arbitrary order.
for x in a.intersection(&b) {
    println!("{x}");
}
let intersection: Set<_, 3> = a.intersection(&b).copied().collect();
assert_eq!(intersection, Set::from([2, 3]));

impl<T, const N: usize> Set<T, N>

pub fn iter(&self) -> Iter<'_, T>

An iterator visiting all elements in arbitrary order. The iterator element type is &'a T.

Examples

use micromap::Set;
let mut set: Set<_, 3> = Set::new();
set.insert("a");
set.insert("b");
// Will print in an arbitrary order.
for x in set.iter() {
    println!("{x}");
}

Performance

In the current implementation, iterating over set takes O(len) time.

impl<T, const N: usize> Set<T, N>

pub const fn capacity(&self) -> usize

Returns the number of elements the set can hold.

Examples

use micromap::Set;
let set: Set<i32, 100> = Set::new();
assert_eq!(set.capacity(), 100);

pub const fn is_empty(&self) -> bool

Returns true if the set contains no elements.

Examples

use micromap::Set;
let mut set: Set<_, 3> = Set::new();
assert!(set.is_empty());
set.insert(1);
assert!(!set.is_empty());

pub const fn len(&self) -> usize

Returns the number of elements in the set.

Examples

use micromap::Set;
let mut set: Set<_, 3> = Set::new();
assert_eq!(set.len(), 0);
set.insert(1);
assert_eq!(set.len(), 1);

pub fn clear(&mut self)

Clears the set, removing all values.

Examples

use micromap::Set;
let mut set: Set<_, 3> = Set::new();
set.insert(1);
set.clear();
assert!(set.is_empty());

pub fn retain<F: FnMut(&T) -> bool>(&mut self, f: F)

Retains only the elements specified by the predicate.

In other words, remove all elements e for which f(&e) returns false. The elements are visited in unsorted (and unspecified) order.

Examples

use micromap::Set;
let mut set = Set::from([1, 2, 3, 4, 5, 6]);
set.retain(|&k| k % 2 == 0);
assert_eq!(set, Set::from([2, 4, 6]));

Performance

In the current implementation, this operation takes O(len) time.

impl<T: PartialEq, const N: usize> Set<T, N>

pub fn contains<Q>(&self, k: &Q) -> bool
where T: Borrow<Q>, Q: PartialEq + ?Sized,

Returns true if the set contains a value.

The value may be any borrowed form of the set’s value type, but PartialEq on the borrowed form must match those for the value type.

Examples

use micromap::Set;
let set = Set::from([1, 2, 3]);
assert_eq!(set.contains(&1), true);
assert_eq!(set.contains(&4), false);

pub fn remove<Q>(&mut self, k: &Q) -> bool
where T: Borrow<Q>, Q: PartialEq + ?Sized,

Removes a value from the set. Returns whether the value was present in the set.

The value may be any borrowed form of the set’s value type, but PartialEq on the borrowed form must match those for the value type.

Examples

use micromap::Set;
let mut set: Set<_, 3> = Set::new();
set.insert(2);
assert_eq!(set.remove(&2), true);
assert_eq!(set.remove(&2), false);

pub fn insert(&mut self, k: T) -> bool

Adds a value to the set. (will not update)

Returns whether the value was newly inserted. That is:

If the set did not previously contain this value, true is returned.

If the set already contained this value, false is returned, and the set is not modified: original value is not replaced, and the value passed as argument is dropped.

Examples

use micromap::Set;
let mut set: Set<_, 3> = Set::new();
assert_eq!(set.insert(2), true);
assert_eq!(set.insert(2), false);
assert_eq!(set.len(), 1);

Panics

It may panic if there are too many items in the set already to contain another new item.

pub fn checked_insert(&mut self, k: T) -> Option<T>

Attempt to insert a value into the set. (will not update, and no panic)

If the value exists, whether the set is full or not, we update the value (exclude key), and return Some(Some(old_value));

If the key does not exist and the set is full already, we can do nothing, so just return None;

If the key does not exist and the set has empty slot, we insert into that slot and return Some(None).

If the value existed, we do nothing, and return Some(T), which is the value parameter we passed in.

If the value does not exist and the set is full already, return Some(T) like above.

If the value does not exist and the set has empty slot, we insert into that slot and return None.

Examples

use micromap::Set;
// Only the first element in the tuple is considered when determining whether
// two `Foo`s are equal.
#[derive(Debug)]
struct Foo(u8, char);
impl PartialEq for Foo {
   fn eq(&self, other: &Self) -> bool {
      self.0 == other.0
    }
}
// Not necessary
impl Eq for Foo {}
// create Set for Foo
let mut set: Set<Foo, 3> = Set::new();
// For `Some(None)`, `Some(_)` indicates that the insertion was successful, `None`
// in the former means that the inserted value was not in set before, that is, the
// insertion occurred, otherwise, no-op was performed.
assert_eq!(set.checked_insert(Foo(1, 'a')), None);
let owned_foo = Foo(1, 'A'); // no `Clone` and `Copy` trait
if let Some(getback_foo) = set.checked_insert(owned_foo) { // take the ownership
    assert_eq!(getback_foo.1, 'A'); // get back the `owned_foo`
} else {
    unreachable!();
}
assert_eq!(set.checked_insert(Foo(2, 'b')), None);
assert_eq!(set.checked_insert(Foo(3, 'c')), None);
assert_eq!(set.len(), set.capacity());
// set is full now.
assert_eq!(set.checked_insert(Foo(2, 'B')).unwrap().1, 'B');
// This insertion will cause capacity overflow, so no insertion is performed
// and `None` is returned.
assert_eq!(set.checked_insert(Foo(4, 'd')), None);

pub unsafe fn insert_unchecked(&mut self, k: T) -> bool

Insert a value into the set without bound check in release mode. (with panic and undefined behavior possible)

Returns whether the value was newly inserted. That is:

If the set did not previously contain this value, true is returned.

If the set already contained this value, false is returned, and the set is not modified: original value is not replaced, and the value passed as argument is dropped.

Examples

use micromap::Set;
let mut set: Set<_, 3> = Set::new();
assert_eq!(set.len(), 0);
assert_eq!(set.capacity(), 3);
unsafe {
    assert_eq!(set.insert_unchecked(0), true);
    assert_eq!(set.insert_unchecked(1), true);
    assert_eq!(set.insert_unchecked(2), true);
    assert_eq!(set.insert_unchecked(2), false);
    assert_eq!(set.len(), set.capacity()); // 3
    // assert_eq!(set.insert_unchecked(3), true); // CAN NOT DO THIS!
}

Panics

It may panic if there are too many items in the set already. Pay attention, it panics only in the debug mode. In the release mode, you are going to get undefined behavior. This is done for the sake of performance, in order to avoid a repetitive check for the boundary condition on every insert().

Safety

Calling this method to add a new key-value pair when the Set is already full is undefined behavior instead of panic. So you need to make sure that the set is not full before calling.

pub fn get<Q>(&self, k: &Q) -> Option<&T>
where T: Borrow<Q>, Q: PartialEq + ?Sized,

Returns a reference to the value in the set, if any, that is equal to the given value.

The value may be any borrowed form of the set’s value type, but PartialEq on the borrowed form must match those for the value type.

Examples

use micromap::Set;
let set = Set::from([1, 2, 3]);
assert_eq!(set.get(&2), Some(&2));
assert_eq!(set.get(&4), None);

pub fn take<Q>(&mut self, k: &Q) -> Option<T>
where T: Borrow<Q>, Q: PartialEq + ?Sized,

Removes and returns the value in the set, if any, that is equal to the given one.

The value may be any borrowed form of the set’s value type, but PartialEq on the borrowed form must match those for the value type.

Examples

use micromap::Set;
let mut set = Set::from([1, 2, 3]);
assert_eq!(set.take(&2), Some(2));
assert_eq!(set.take(&2), None);

impl<T: PartialEq, const N: usize> Set<T, N>

Specialized methods available only on Set.

pub fn is_disjoint<const M: usize>(&self, other: &Set<T, M>) -> bool

Returns true if self has no elements in common with other. This is equivalent to checking for an empty intersection.

Examples

use micromap::Set;
let a = Set::from([1, 2, 3]);
let mut b: Set<u32, 5> = Set::new();
assert_eq!(a.is_disjoint(&b), true);
b.insert(4);
assert_eq!(a.is_disjoint(&b), true);
b.insert(1);
assert_eq!(a.is_disjoint(&b), false);

pub fn is_subset<const M: usize>(&self, other: &Set<T, M>) -> bool

Returns true if the set is a subset of another, i.e., other contains at least all the values in self.

Examples

use micromap::Set;
let sup = Set::from([1, 2, 3]);
let mut set: Set<u32, 5> = Set::new();
assert_eq!(set.is_subset(&sup), true);
set.insert(2);
assert_eq!(set.is_subset(&sup), true);
set.insert(4);
assert_eq!(set.is_subset(&sup), false);

pub fn is_superset<const M: usize>(&self, other: &Set<T, M>) -> bool

Returns true if the set is a superset of another, i.e., self contains at least all the values in other.

Examples

use micromap::Set;
let sub = Set::from([1, 2]);
let mut set: Set<u32, 5> = Set::new();
assert_eq!(set.is_superset(&sub), false);
set.insert(0);
set.insert(1);
assert_eq!(set.is_superset(&sub), false);
set.insert(2);
assert_eq!(set.is_superset(&sub), true);

pub fn replace(&mut self, value: T) -> Option<T>

Adds a value to the set, replacing the existing value, if any, that is equal to the given one. Returns the replaced value.

Examples

use micromap::Set;
let mut set: Set<_, 5> = Set::new();
set.insert(Vec::<i32>::new());
assert_eq!(set.get(&[][..]).unwrap().capacity(), 0);
set.replace(Vec::with_capacity(10));
assert_eq!(set.get(&[][..]).unwrap().capacity(), 10);

impl<T: PartialEq, const N: usize> Set<T, N>

pub fn symmetric_difference<'a, const M: usize>( &'a self, other: &'a Set<T, M>, ) -> SymmetricDifference<'a, T, N, M>

Visits the values representing the symmetric difference, i.e., the values that are in self or other but not in both.

Examples

use micromap::Set;
let a = Set::from([1, 2, 3]);
let b = Set::from([4, 2, 3, 4]);
// Can be seen as `(a - b) ∪ (b - a)`.
for x in a.symmetric_difference(&b) {
    println!("{x}"); // Print 1, 4
}
let sym_diff: Set<_, 7> = a.symmetric_difference(&b).copied().collect();
assert_eq!(sym_diff, Set::from([1, 4]));

impl<T: PartialEq, const N: usize> Set<T, N>

pub fn union<'a, const M: usize>( &'a self, other: &'a Set<T, M>, ) -> Union<'a, T, M>

Visits the values representing the union, i.e., the values that are in self or other.

Examples

use micromap::Set;
let a = Set::from([1, 2, 3]);
let b = Set::from([4, 2, 3, 4]);
// Can be seen as `a ∪ b`.
for x in a.union(&b) {
    println!("{x}"); // Print 1, 2, 3, 4
}
let union: Set<_, 7> = a.union(&b).copied().collect();
assert_eq!(union, Set::from([1, 2, 3, 4]));